migrate from truffle to hardhat & add gas reporter

+++ contracts/test_token_erc721.sol
@@ -5,7 +5,7 @@ pragma solidity >0.4.22;
 import "openzeppelin-solidity/contracts/token/ERC721/ERC721.sol";
 
 contract TestToken_721 is ERC721 {
-    constructor(uint256 initialSupply) ERC721("TestToken_721", "TEST") public{
+    constructor(uint initialSupply) ERC721("TestToken_721", "TEST") public{
         for (uint i=0; i<initialSupply; i++) {
           _mint(msg.sender, i);
         }
